src/processor/requirements.txt
nats-py==2.7.0
psycopg2-binary==2.9.9
scikit-learn==1.4.2
numpy==1.26.4
